requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
dnspython==1.11.1
gunicorn==18.0
itsdangerous==0.23
redis==2.9.0
wsgiref==0.1.2
